+++ packages/cli/src/lib/cli.spec.ts
@@ -1,17 +1,20 @@
+import { Effect } from 'effect';
 import { cli } from './cli';
 
 describe('cli', () => {
-  it('should work', () => {
-    expect(
-      () => cli([
-        '', // node
-        '', // javascript filename
-        'query',
-        '-W',
-        'topsecret',
-        '-c',
-        'select current_timestamp; select 1',
-      ]),
-    ).not.toThrow();
+  it('should work', async () => {
+    await expect(
+      Effect.runPromise(
+        cli([
+          '', // node
+          '', // javascript filename
+          '-W',
+          'topsecret',
+          'query',
+          '-c',
+          'select current_timestamp; select 1',
+        ]),
+      ),
+    ).resolves.not.toThrow();
   });
 });