+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jmorecroft67/pg-stream-cli
 
+## 2.1.0
+
+### Minor Changes
+
+- new recvlogical subcommand for logical streaming to table (default) or JSON, output per transaction
+- query output now selectable, table (default) or JSON
+- improved help
+
 ## 2.0.1
 
 ### Patch Changes